Push flows to ntopng
Ntopng is a an open-source traffic analyser:
https://www.ntop.org/products/traffic-analysis/ntop/

Unfortunately, it cannot ingest Netflow data directly, instead,
it requires a payware component (Nprobe) to tranlate Netflow to JSON
encapsulated in ZeroMQ messages.

This patch allows softflowd to generate those messages directly,
allowing Ntopng to be used without the need for the payware component.

Signed-off-by: Alastair D'Silva <alastair@d-silva.org>